Devide study_tests
